move remove code to command line
